gdm: Fix misuse of multi-user.target
We should still be able to login on tty when we ask systemd to stop at
multi-user.target (usually via kernel command line: adding a "3" or
systemd.unit=multi-user.target).

After this patch this can be done and one can even boot on
multi-user.target and later ask systemctl to isolate graphical.target,
switching to it.

Another thing that this fixes is the ugly effect of starting getty on
tty1 and then switching to tty2 for the graphical login. Now gdm.service
will conflict with getty@tty1.service so when booting into
graphical.target getty@tty1 is not even started. Since it conflicts,
it's still possible to switch to it later since getty will be stopped
and gdm takes it over.

End result: when booting on graphical.target gdm will run on tty1 -
after login the session stays on tty1 and the greeter runs on tty7.
Text console starts on tty2 via autovt feature from systemd.